requirements.txt
lxml==4.2.5
numpy==1.15.4
pkg-resources==0.0.0
python-Levenshtein==0.12.0
Unidecode==1.0.23
